Move Sample to timeseries._base_types
This is to allow adding some other base types in future commits.
