+++ src/frequenz/sdk/timeseries/__init__.py
@@ -10,8 +10,8 @@ A timeseries is a stream (normally an async iterator) of
 This module provides tools to operate on timeseries.
 """
 
+from ._base_types import Sample
 from ._resampler import GroupResampler, Resampler, ResamplingFunction
-from ._sample import Sample
 
 __all__ = [
     "GroupResampler",
