+++ src/frequenz/sdk/timeseries/_base_types.py
@@ -1,7 +1,7 @@
 # License: MIT
 # Copyright © 2022 Frequenz Energy-as-a-Service GmbH
 
-"""Common types for the Data Pipeline."""
+"""Timeseries basic types."""
 
 from dataclasses import dataclass
 from datetime import datetime
